pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaoyan"
version = "0.1.0"
description = "小研 — 基于飞书的私人科研助理"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "XiaoYan Contributors"}]

dependencies = [
    "python-dotenv>=1.0",
    "openai>=1.0",
    "apscheduler>=3.10,<4",
    "playwright>=1.40",
    "pyperclip>=1.8",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]
bibtex = [
    "bibtexparser>=1.4",
]

[project.scripts]
xiaoyan = "xiaoyan.cli:main"

[tool.setuptools.packages.find]
include = ["xiaoyan*"]
